requirements.txt
google-cloud-secret-manager
gspread
oauth2client